Collect data average
This commit adds one more map, now using a callable class.
The callable class is responsible for keeping track of the last n values received from the stream and calculating their average value.
A helper function "getReadingValue" was imported so we can easily access the heart rate value of tuples we receive.

Signed-off-by: Gabriel R. Sezefredo <gabriel@sezefredo.com.br>